feat(notification): add outlined to banner and inline
