+++ src/components/Notification/InlineNotification/InlineNotification.tsx
@@ -3,7 +3,7 @@
 import { jsx } from '@emotion/core';
 import * as React from 'react';
 import CompactNotification from '../subcomponents/CompactNotification';
-import { NotificationTypes } from '../Notification';
+import { NotificationStyleType, NotificationTypes } from '../Notification';
 import { TestId } from '../../../utils/types';
 
 export type Props = {
@@ -13,6 +13,8 @@ export type Props = {
   message: string;
   /** The type of the Notification */
   type: NotificationTypes;
+  /** The style type of the Notification. Defaults to elevated */
+  styleType: NotificationStyleType;
   /** The primary call-to-action label of the Notification */
   primaryCTALabel?: string;
   /** The primary call-to-action of the Notification */
@@ -27,6 +29,7 @@ const InlineNotification: React.FC<Props> = ({
   withIcon = false,
   message,
   type,
+  styleType = 'elevated',
   primaryCTALabel,
   primaryCTA = undefined,
   closeCTA = undefined,
@@ -38,6 +41,7 @@ const InlineNotification: React.FC<Props> = ({
       message={message}
       variant="inline"
       type={type}
+      styleType={styleType}
       primaryCTALabel={primaryCTALabel}
       primaryCTA={primaryCTA}
       closeCTA={closeCTA}
