+++ src/components/Notification/Notification.tsx
@@ -1,3 +1,5 @@
 export type NotificationTypes = 'success' | 'error' | 'info' | 'warning';
 
 export type NotificationVariants = 'inline' | 'banner';
+
+export type NotificationStyleType = 'elevated' | 'outlined';
